Viewpager moved to new file
